Reorganize Rust code
