--- rust/types/mod.rs
@@ -1,3 +1,0 @@
-mod tzinfo;
-
-pub use tzinfo::TzInfo;
