--- rust/types/tzinfo.rs
@@ -1,40 +1,0 @@
-use pyo3::prelude::*;
-use pyo3::types::{PyDateTime, PyDelta, PyTzInfo};
-
-#[pyclass(module = "_pendulum", extends = PyTzInfo)]
-pub struct TzInfo {
-    seconds: i32,
-}
-
-#[pymethods]
-impl TzInfo {
-    #[new]
-    pub fn new(seconds: i32) -> Self {
-        Self { seconds }
-    }
-
-    fn utcoffset<'p>(&self, py: Python<'p>, _dt: &PyDateTime) -> PyResult<&'p PyDelta> {
-        PyDelta::new(py, 0, self.seconds, 0, true)
-    }
-
-    fn tzname(&self, _dt: &PyDateTime) -> String {
-        self.__str__()
-    }
-
-    fn dst(&self, _dt: &PyDateTime) -> Option<&PyDelta> {
-        None
-    }
-
-    fn __repr__(&self) -> String {
-        format!("TzInfo({})", self.__str__())
-    }
-
-    fn __str__(&self) -> String {
-        if self.seconds == 0 {
-            "UTC".to_string()
-        } else {
-            let mins = self.seconds / 60;
-            format!("{:+03}:{:02}", mins / 60, (mins % 60).abs())
-        }
-    }
-}
